+++ CHANGELOG.md
@@ -2,14 +2,65 @@
 
 | モジュール/プラグイン名 | Description | 最新のバージョン |
 | :-- | :-- |:---------|
-| core | イベントトラッキング機能を提供します。 | 2.32.0   |
-| inappmessaging | アプリ内メッセージ機能を提供します。 | 2.25.0   |
-| notifications | プッシュ通知の受信および効果測定機能を提供します。 | 2.16.0   |
-| variables | 設定値配信機能を提供します。 | 2.11.0   |
-| visualtracking | ビジュアルトラッキング機能を提供します。| 2.13.0   |
-| inappframe          | アプリ内の特定の場所に任意の表示ができます     | 0.6.0    |
-| Karte Gradle Plugin | ビジュアルトラッキング機能に必要なプラグインです。| 2.7.0    |
-| debugger         | イベントをわかりやすく管理画面で確認できます。 | 1.2.0    |
+| core | イベントトラッキング機能を提供します。 | 2.33.0   |
+| inappmessaging | アプリ内メッセージ機能を提供します。 | 2.26.0   |
+| notifications | プッシュ通知の受信および効果測定機能を提供します。 | 2.17.0   |
+| variables | 設定値配信機能を提供します。 | 2.12.0   |
+| visualtracking | ビジュアルトラッキング機能を提供します。| 2.14.0   |
+| inappframe          | アプリ内の特定の場所に任意の表示ができます     | 0.7.0    |
+| Karte Gradle Plugin | ビジュアルトラッキング機能に必要なプラグインです。| 2.8.0    |
+| debugger         | イベントをわかりやすく管理画面で確認できます。 | 1.3.0    |
+
+# Releases - 2026.01.14
+
+### Core 2.33.0
+**🔨CHANGED**
+- 依存ライブラリをアップデートしました。
+- Deprecatedとなったandroidx.ads:ads-identifierライブラリの使用箇所を削除しました。
+- Version Catalogに対応しました。
+- targetSdkを36にしました。
+
+### InAppMessaging 2.26.0
+**🔨CHANGED**
+- 依存ライブラリをアップデートしました。
+- Version Catalogに対応しました。
+- targetSdkを36にしました。
+
+### Notifications 2.17.0
+**🔨CHANGED**
+- 依存ライブラリをアップデートしました。
+- Version Catalogに対応しました。
+- targetSdkを36にしました。
+
+### Variables 2.12.0
+**🔨CHANGED**
+- 依存ライブラリをアップデートしました。
+- Version Catalogに対応しました。
+- targetSdkを36にしました。
+
+### VisualTracking 2.14.0
+**🔨CHANGED**
+- 依存ライブラリをアップデートしました。
+- Version Catalogに対応しました。
+- targetSdkを36にしました。
+
+### InAppFrame 0.7.0
+**🔨CHANGED**
+- 依存ライブラリをアップデートしました。
+- Version Catalogに対応しました。
+- targetSdkを36にしました。
+
+### Karte Gradle Plugin 2.8.0
+**🔨CHANGED**
+- 依存ライブラリをアップデートしました。
+- Version Catalogに対応しました。
+- targetSdkを36にしました。
+
+### Debugger 1.3.0
+**🔨CHANGED**
+- 依存ライブラリをアップデートしました。
+- Version Catalogに対応しました。
+- targetSdkを36にしました。
 
 # Releases - 2025.12.03
 
@@ -65,7 +116,7 @@
 
 # Releases - 2025.07.14
 
-### core 2.31.0
+### Core 2.31.0
 **🎉FEATURE**
 - WebViewの情報を送るようになりました。
   詳細は [イベントに自動追加されるフィールド](doc:appendix-fields-android-sdk-v2) をご覧ください。
@@ -228,7 +279,7 @@
 **💊FIXED**
 - _message_suppressed送信前に、まれにクラッシュする不具合を修正しました。
 
-### notifications 2.12.0
+### Notifications 2.12.0
 **💊FIXED**
 - タスクキル状態でPush通知を開封した際の不具合を修正しました。
 
@@ -697,6 +748,7 @@
 - SDK内部で参照しているライブラリ名が難読化時に意図しない形式になってしまう不具合を修正しました。
 
 # Releases - 2020.09.16
+
 ### Core 2.5.0
 **🔨CHANGED**
 - サポート改善のため、SDKのログを一時保持・収集するようにしました。
@@ -711,11 +763,13 @@
 - SDK内部で参照しているライブラリ名を変更しました。
 
 # Releases - 2020.08.21
+
 ### InAppMessaging 2.5.1
 **💊FIXED**
 - 接客表示中の画面遷移時やviewイベント発火時に、クラッシュする可能性がある不具合を修正しました。([issue](https://github.com/plaidev/karte-android-sdk/issues/1))
 
 # Releases - 2020.07.22
+
 ### InAppMessaging 2.5.0
 **🎉FEATURE**
 - アクションを常駐させるオプションに対応しました。
@@ -730,6 +784,7 @@
 - Java9以降のモジュール機能に対応したライブラリを参照するとビルドができない不具合を修正しました。
 
 # Releases - 2020.07.08
+
 ### Core 2.4.0
 **🎉FEATURE**
 - KARTE固有のURLスキームからNative機能の呼び出しが出来るようになりました。
@@ -744,6 +799,7 @@
 - Coreの機能追加に伴う内部処理の変更を行いました。
 
 # Releases - 2020.07.06
+
 ### Core 2.3.2
 **💊FIXED**
 - SDK初期化時にネットワーク状態の切り替えが起こると、クラッシュする可能性がある不具合を修正しました。([issue](https://github.com/plaidev/karte-android-sdk/issues/2))
@@ -753,11 +809,13 @@
 - Android Gradle Plugin 4.0.0を使用しているとビルドできない問題を修正しました。
 
 # Releases - 2020.06.25
+
 ### Core 2.3.1
 **💊FIXED**
 - AndroidManifestにandroid:sharedUserIdを指定した際にライブラリがロードできなくなる問題を修正しました。
 
 # Releases - 2020.06.23
+
 ### Core 2.3.0
 **🎉FEATURE**
 - ディープリンクによるアプリ流入時に自動で送信するイベントを追加しました。
@@ -772,6 +830,7 @@
 
 
 # Releases - 2020.05.29
+
 ### Core 2.2.0
 **💊FIXED**
 - SDK初期化時にネットワーク状態の切り替えが起こると、クラッシュする可能性がある不具合を修正しました。
@@ -790,6 +849,7 @@
 - trackOpen / trackClick のAPIにJSONObjectを受け入れるインターフェースを追加しました。
 
 # Releases - 2020.04.24
+
 ### Core 2.1.0
 **🔨CHANGED**
 - イベント送信時のリクエストボディをgzip圧縮するよう変更しました。
@@ -809,8 +869,8 @@
 **🔨CHANGED**
 - Androidxへの実行時の依存を廃止しました。
 
-
 # Releases - 2020.04.07
+
 ### Core 2.0.0
 **🎉FEATURE**
 - イベントの送信失敗時に再送が行われるようになりました。